requirements.txt
Keras==2.2.4
Pillow==5.3.0
flickrapi==2.4.0
